feat: adding controls over destructive endpoints and better db seeding

+++ api_views/main.py
@@ -1,15 +1,20 @@
 from flask import Response
-
+import sys
 from models.user_model import *
+from app import destructive
+from api_views.users import error_message_helper
 
 
 def populate_db():
-    db.drop_all()
-    db.create_all()
-    User.init_db_users()
-    response_text = '{ "message": "Database populated." }'
-    response = Response(response_text, 200, mimetype='application/json')
-    return response
+    if destructive:
+        db.drop_all()
+        db.create_all()
+        User.init_db_users()
+        response_text = '{ "message": "Database populated." }'
+        response = Response(response_text, 200, mimetype='application/json')
+        return response
+    else:
+        return Response(error_message_helper("Destructive actions not allowed"), 401, mimetype="application/json")
 
 
 def basic():
